React is garbage
Add getStickyUntil method to determine the slide index from a name reference. This map of names is updated when children are updated (to account for async yaml parsing). I also took the liberty of renaming routes that were hardcoded to hide the header/footer and just prefixed them with /slides, change it if you disagree. The DocumentTitle component is used and documentation includes use of `name` in quick slides.